% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{epts_example}
\alias{epts_example}
\title{Get path to epts example}
\usage{
epts_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
epts comes bundled with example files in its \code{inst/extdata}
directory. This function make them easy to access.
}
\examples{
epts_example()
epts_example("fifa_example.txt")
}
