% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{forestPlotMultiArms}
\alias{forestPlotMultiArms}
\title{Internal plotting function for comparison plots}
\usage{
forestPlotMultiArms(
  eefAnalyticsList,
  group,
  Conditional = TRUE,
  ES_Total = TRUE,
  modelNames,
  intlabels = NULL,
  intcolors = NULL,
  maintitle = NULL,
  xlabel = NULL,
  ylabel = NULL,
  vlinecolor = "black"
)
}
\description{
Internal plotting function for comparison plots
}
\keyword{internal}
