% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_epts}
\alias{read_epts}
\title{Read EPTS data}
\usage{
read_epts(data, metadata)
}
\arguments{
\item{data}{Path to a raw data txt file.}

\item{metadata}{Path to a format specification xml file.}
}
\value{
A data frame containing a representation of the EPTS data.
}
\description{
Read EPTS data
}
\examples{
data <- epts_example("fifa_example.txt")
metadata <- epts_example("fifa_example.xml")
read_epts(data, metadata)
}
