% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEPTS.R
\name{runEPTS}
\alias{runEPTS}
\title{Run the EPTS Shiny Application}
\usage{
runEPTS()
}
\description{
Launches a Shiny application that provides an interactive user interface
to run the functions provided by the epts package.
}
\examples{
if (interactive()) {
  runEPTS()
}

}
