% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epts-package.R
\docType{package}
\name{epts-package}
\alias{epts}
\alias{epts-package}
\title{epts: Read EPTS Data}
\description{
Import Electronic Performance and Tracking Systems (EPTS) data into R. Reads a raw data file using the data format specification and metadata from a XML-Schema document.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/george-wood/epts}
  \item Report bugs at \url{https://github.com/george-wood/epts/issues}
}

}
\author{
\strong{Maintainer}: George Wood \email{georgebwood@gmail.com} (\href{https://orcid.org/0000-0003-0046-4465}{ORCID})

}
\keyword{internal}
